Add set data type that will be used by proto.
