+++ models/src/main/protobuf/RhoTypes.proto
@@ -6,6 +6,7 @@ option java_package = "coop.rchain.models";
 
 option (scalapb.options) = {
   import: "coop.rchain.models.BitSetBytesMapper.bitSetBytesMapper"
+  import: "coop.rchain.models.rholang.sort.ordering.parOrdering"
 };
 
 message Par {
@@ -150,6 +151,7 @@ message Expr {
         EList e_list_body = 20;
         ETuple e_tuple_body = 21;
         ESet e_set_body = 22;
+//        bytes e_set_body = 22 [(scalapb.field).type = "coop.rchain.models.ParSet"];
         EMap e_map_body = 23;
         EMethod e_method_body = 24;
         Channel e_eval_body = 26;
@@ -170,7 +172,7 @@ message ETuple {
 }
 
 message ESet {
-    repeated Par ps = 1;
+    repeated Par ps = 1 [(scalapb.field).collection_type = "coop.rchain.models.ParTreeSet"];
     bytes locallyFree = 3 [(scalapb.field).type = "scala.collection.immutable.BitSet"];
     bool connective_used = 4;
 }
