+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/RholangCLI.scala
@@ -9,9 +9,10 @@ import cats.{Monad, _}
 import cats.implicits._
 import cats.syntax._
 import coop.rchain.catscontrib.Capture._
+import coop.rchain.models.rholang.sort.ParSortMatcher
 import coop.rchain.models.{BindPattern, Channel, Par, TaggedContinuation}
 import coop.rchain.rholang.interpreter.errors._
-import coop.rchain.rholang.interpreter.implicits.VectorPar
+import coop.rchain.models.rholang.implicits.VectorPar
 import coop.rchain.rholang.interpreter.storage.StoragePrinter
 import coop.rchain.rholang.syntax.rholang_mercury.Absyn.Proc
 import coop.rchain.rholang.syntax.rholang_mercury.{parser, Yylex}
@@ -134,12 +135,10 @@ object RholangCLI {
                                  DebruijnLevelMap[VarSort]())
         outputs <- normalizeTerm[Coeval](term, inputs)
         par <- ParSortMatcher
-                .sortMatch[Coeval](Some(outputs.par))
+                .sortMatch(Some(outputs.par))
                 .map(_.term)
-                .flatMap {
-                  case None      => Coeval.raiseError[Par](SortMatchError("ParSortMatcher failed"))
-                  case Some(par) => Coeval.delay(par)
-                }
+                .leftMap(th => SortMatchError(th.getMessage))
+                .fold(err => Coeval.raiseError[Par](err), par => Coeval.delay(par))
       } yield par
     } catch {
       case th: Throwable => Coeval.raiseError(UnrecognizedInterpreterError(th))
