+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/NormalizeTest.scala
@@ -8,7 +8,7 @@ import coop.rchain.rholang.syntax.rholang_mercury.Absyn.{
   _
 }
 import org.scalatest._
-
+import coop.rchain.models.rholang.sort.ordering._
 import scala.collection.immutable.BitSet
 import coop.rchain.models.Channel.ChannelInstance._
 import coop.rchain.models.Connective.ConnectiveInstance._
@@ -18,7 +18,7 @@ import coop.rchain.models.Var.WildcardMsg
 import coop.rchain.models._
 import coop.rchain.rspace.Serialize
 import errors._
-import implicits._
+import coop.rchain.models.rholang.implicits._
 import monix.eval.Coeval
 import org.scalactic.TripleEqualsSupport
 
@@ -113,12 +113,13 @@ class CollectMatcherSpec extends FlatSpec with Matchers {
 
     val result = ProcNormalizeMatcher.normalizeMatch[Coeval](set, inputs).value
     result.par should be(
-      inputs.par.prepend(
-        ESet(List[Par](EPlus(EVar(BoundVar(1)), EVar(FreeVar(0))),
-                       GInt(7),
-                       GInt(8).prepend(EVar(FreeVar(1)))),
-             locallyFree = BitSet(1),
-             connectiveUsed = true)))
+      inputs.par.prepend(ESet(
+        SortedHashSet(Seq[Par](EPlus(EVar(BoundVar(1)), EVar(FreeVar(0))),
+                               GInt(7),
+                               GInt(8).prepend(EVar(FreeVar(1))))),
+        locallyFree = BitSet(1),
+        connectiveUsed = true
+      )))
     result.knownFree should be(
       inputs.knownFree.newBindings(List(("R", ProcSort, 0, 0), ("Q", ProcSort, 0, 0)))._1)
   }
