+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/ReduceSpec.scala
@@ -12,7 +12,7 @@ import coop.rchain.models.Var.VarInstance._
 import coop.rchain.models.{GPrivate => _, _}
 import coop.rchain.rholang.interpreter.errors._
 import coop.rchain.rholang.interpreter.errors.ReduceError
-import coop.rchain.rholang.interpreter.implicits._
+import coop.rchain.models.rholang.implicits._
 import coop.rchain.rholang.interpreter.storage.implicits._
 import coop.rchain.rspace.internal.{Datum, Row, WaitingContinuation}
 import coop.rchain.rspace.{IStore, LMDBStore, Serialize}
@@ -650,7 +650,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "eval of `toByteArray` method on any process" should "return that process serialized" in {
-    import coop.rchain.models.implicits._
+    import coop.rchain.models.serialization.implicits._
     val proc = Receive(Seq(ReceiveBind(Seq(ChanVar(FreeVar(0))), Quote(GString("channel")))),
                        Par(),
                        false,
@@ -698,7 +698,7 @@ class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   }
 
   "eval of hexToBytes" should "transform encoded string to byte array (not the rholang term)" in {
-    import coop.rchain.models.implicits._
+    import coop.rchain.models.serialization.implicits._
     val testString                = "testing testing"
     val base16Repr                = Base16.encode(testString.getBytes)
     val proc: Par                 = GString(base16Repr)
